zh jrw d qxpehu rqh ylfwrub urbdoh >>=3
Added the code for the __sendImage() and __readAndStoreImageBytes() functions. Also edited some of the calls to the functions in order to work with the updated calls that I did. Also removed a semicolon that I'm not sure should've been there?
Need to update the sendImage and readAndStoreImageBytes functions with the new ones, but those might not work yet lmao